% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concat_rank_value_occurrence.R
\name{concat_rank_value_occurrence}
\alias{concat_rank_value_occurrence}
\title{Wrapper for concatenating rank, value, and number of occurrences.}
\usage{
concat_rank_value_occurrence(r, n, o)
}
\arguments{
\item{r}{something.}

\item{n}{something.}

\item{o}{something.}
}
\value{
a vector.
}
\description{
concat_rank_value_occurrence takes vectors representing rank, value, and number of occurrences and returns them in a concatenated string. This is just a wrapper around paste0().
}
\examples{
# Example
concat_rank_value_occurrence(1, "a", 100)
}

