% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequent_values.R
\name{frequent_values}
\alias{frequent_values}
\title{Find the least or most frequent occuring values in an atomic vector..}
\usage{
frequent_values(.x, n = 5L, type = "most")
}
\arguments{
\item{.x}{something.}

\item{n}{something.}

\item{type}{something.}
}
\value{
a vector.
}
\description{
frequent_values takes a vector and returns the n least or most occuring values.
}

