% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineprof.R
\name{lineprof}
\alias{lineprof}
\title{Identifies significant changes in a changing data set with time}
\usage{
lineprof(data, L_per_pix = 1, type = "meanvar")
}
\arguments{
\item{data}{a dataframe by read.csv of a .csv created from using the ImageJ macro found in vignette. Or simply a dataframe with each column representing a new instance (with potential changes) of a data set of same N}

\item{L_per_pix}{a unit conversion of length per pixel, or desired unit per data unit default = 1 or no conversion}

\item{type}{type of changepoint desired. Choose between "meanvar" (mean and variance), "mean", "variance", and "all" (changepoints for meanvar, mean, and variance); note the default is "meanvar" see Killick et al., 2013 for more details}
}
\value{
results in a dataframe with time interval, filename, changepoint in units of original data, a conversion labeled as unit.length
}
\description{
\code{lineprof} identifies the most significant statistical change in mean, variance, or both mean and variance in the data (often pixel data) as specified by user on changing datasets
}
\examples{
changepoint_df <- lineprof(data = df, L_per_pix = 0.123, type = "mean")
changepoint_df <- lineprof(data = df, L_per_pix = 500, type = "variance")
}
