% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\alias{profile.character}
\alias{profile.data.frame}
\alias{profile.default}
\alias{profile.numeric}
\title{Profile atomic vectors or data.frames.}
\usage{
profile(.x)

\method{profile}{numeric}(.x)

\method{profile}{character}(.x)

\method{profile}{data.frame}(.x)

\method{profile}{default}(.x)
}
\arguments{
\item{.x}{a vector or data.frame to be profiled.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{profile} takes vectors and data.frames and returns a data.frame containing important descriptive statistics.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Method for numeric.

\item \code{character}: Method for character.

\item \code{data.frame}: Method for data.frames.

\item \code{default}: Method for default.
}}
\examples{
# Example
profile(mtcars)
profile(iris)
profile(state.name)
}

