% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_numeric}
\alias{profile_numeric}
\title{Profile numeric atomic vectors.}
\usage{
profile_numeric(.x)
}
\arguments{
\item{.x}{a numeric atomic vector or data.frame to be profiled.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{profile_numeric} takes numeric vectors and returns a data.frame containing important descriptive statistics.
}
\examples{
# Example
profile_numeric(1:100)
}

